Rename extensions to defaults
