+++ src/importnb/__init__.py
@@ -6,6 +6,6 @@ __all__ = "Notebook", "reload", "MAIN", "CLI", "INTERACTIVE", "IMPORTED", "Remot
 from .loader import Notebook, reload, unload_ipython_extension
 from .parameterize import parameterize, Parameterize
 from .remote import Remote
-from .extensions import load_ipython_extension
+from .defaults import load_ipython_extension, unload_ipython_extension
 from ._version import *
 from .helpers import MAIN, CLI, INTERACTIVE, IMPORTED
